move dataset helpers directly into dataset mutation sub-module
